pyproject.toml
[project]
name = "imgmatch-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Tuấn Khang",email = "nguyenhuutuankhang412@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.1.0,<12.0.0)",
    "dlib (>=19.24.6,<20.0.0)",
    "face-recognition (>=1.3.0,<2.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "imgmatch_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn_runner:run"

